pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aiir"
version = "0.1.0"
description = "AI-powered Incident Response analysis toolset"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "openai>=1.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "rich>=13.0",
    "keyring>=25.0",
]

[project.scripts]
aiir = "aiir.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/aiir"]

[dependency-groups]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-mock>=3.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-mock>=3.0",
]
